instnotemapedit

+++ trunk/cpsycle/host/src/machineui.c
@@ -209,7 +209,7 @@ psy_ui_Component* machineui_create(psy_audio_Machine* machine,
 
 void machineui_drawhighlight(psy_ui_Graphics* g, psy_ui_RealRectangle position)
 {
-	static intptr_t d = 5; // the distance of the highlight from the machine
+	static double d = 5.0; // the distance of the highlight from the machine
 	static psy_ui_RealPoint dirs[] = {
 		{ 1.0, 0.0 }, { 0.0, 1.0 }, { -1.0, 0.0 }, { 0.0, -1.0 }
 	};
